the lastest dependencies and configuration changes

+++ mobile-money-engine/MM-engine/src/main/java/org/mifos/mmoney/models/Configurations.java
@@ -1,11 +1,11 @@
 package org.mifos.mmoney.models;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name="configurations")
@@ -15,7 +15,8 @@ public class Configurations {
 	@Id
 	@Column(name="id")
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	// private int id;
+	private Long id;
 	
 	// Constructor
 	public Configurations(){}
@@ -60,7 +61,10 @@ public class Configurations {
 		this.orgAccId = orgAccId;
 	}
 
-	public int getId() {
+	// public int getId() {
+	// 	return id;
+	// }
+	public Long getId() {
 		return id;
 	}
